split production/dev webpack config
